# Speedup for rcd demosaic