… rtengine/demosaic_algos.cc