@heckflosse 2018-02-28
With last commit I reduced the number of calculations in `STEP 1: Find cardinal and diagonal interpolation directions`

```
                     before     first commit   last commit

100 MP              3850 ms        1320 ms       1140 ms
24 MP                940 ms         340 ms        295 ms
```