@heckflosse 2018-02-28
Processing times with `another small speedup`

```
100 MP : 1090 ms
24 MP:    285 ms
```