@heckflosse 2018-03-02
Processing times with last commit

```
100 MP : 1085 ms
24 MP:    270 ms
```